+++ blackfynn/api/data.py
@@ -201,6 +201,7 @@ class PackagesAPI(APIBase):
         """
         d = pkg.as_dict()
         d.update(kwargs)
+        d.pop("state",None)
         resp = self._put(self._uri('/{id}', id=pkg.id), json=d)
         pkg = self._get_package_from_data(resp)
         return pkg