+++ CHANGELOG.md
@@ -6,6 +6,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Documentation at https://developer.blackfynn.io/python/
 
+## [2.14.0]
+
+### Fixed
+- Adding and retrieving array model values
+
+### Changed
+- Deprecated the `set_type` method on model properties. You can set property types using `ModelProperty.type` directly.
+- Deprecated the `set_value` method on model values. You can now set property values using `ModelValue.value` directly.
+
+## [2.13.1]
+### Fixed
+- Incorrect version
+
+## Added
+- Documentation about release
+
+## [2.13.0]
+### Removed
+- State of packages can no longer be set through this library
+
+## [2.12.0]
+### Removed
+- Separate host for Models API
+
 ## [2.11.1]
 ### Changed
 - Using relate_to on concepts now populates the source and destination of the relationship